requirements.txt
# streamlit==1.29.0
# requests==2.31.0
# pandas==2.1.3
# plotly==5.18.0
# python-dotenv==1.0.0
# streamlit-option-menu==0.3.6
# streamlit-extras==0.3.6
# psycopg2-binary==2.9.6
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
psycopg2-binary==2.9.9
openai==1.51.0
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
pydantic==2.9.2